requirements.txt
gradio>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
dspy-ai>=2.4.0
pydantic>=2.5.0
boto3>=1.34.0
unstructured[pdf]>=0.11.0
python-multipart>=0.0.6
pandas>=2.1.0
plotly>=5.17.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.0
pillow>=10.1.0